fix: use classifier in package name
When two dependencies with the same groupId and artifiactId appear in the
dep tree, but they have a different classifier we should treat them as
different nodes in the tree. At the moment we overwrite these nodes,
which can lead to incorrect results form snyk test.

Adding new test for duplicate dep with classifier.
Fixing parse-mvn to correctly identify each part of a dependency.