feat: produce flat dep-graph scanning unmanaged
Instead of creating a temporary pom file to test unmanaged files, construct a dep-graph from the results of searching search.maven.org.

When using a temporary pom files and scanning multiple unmanaged files, separate transitive lines can interfer with one another, which is not the intension of the feature.

We want to report vulnerabilities in archives found on disk. There is no need to walk transitive lines. So construct a flat dep-graph where each resolved archive is a top-level dependency in that graph.

Removing 'tmp' dependency which is no longer required.

Returning early when scanning archives and no longer shelling out and running maven command line.