refactor: parse digraph output
Implement new functions and class to parse maven digraph output.

Introduces new interface 'MavenGraph' and class 'MavenGraphBuilder'.
To better capture the text being parsed.
We construct a 'MavenGraph' which captures each node and what it dependsOn.
This will be converted to a Snyk dep-graph in later PRs.

Tests document the boundaries between functions and modules.