feat: maven aggregate project option
Introduce a new option to test a project using one maven command but return multiple scan results.

If option 'mavenAggregateProject' is true, perform 'mvn test-compile dependency:tree -DoutputType=dot'.
This command better supports more complex Maven projects where modules depend on one another and are not discoverable in either a local or remote repositories.
In this mode we rely on the Maven reactor resolving inter module dependencies and reporting multiple digraphs on stdout which we parse into a MavenGraph and then into a Snyk dep-graph.

Fixes:
* https://github.com/snyk/snyk-mvn-plugin/issues/124
* https://github.com/snyk/snyk-mvn-plugin/issues/38