feat: Keep `test` scoped deps that transitively reach prod-like deps
This is a workaround to deal with how Maven dependency:tree outputs
dependency information. Prod-like dependencies are not shown again if
first seen via a test scoped dependency.

This change ensures that we keep any prod-like dependencies that are
first seen through a test scope dependency.

A side-effect is that some `test` scoped dependencies will be included
in the dep-graph even when `test` scoped deps are explicitly excluded.